code.cleanup: refactor QueryService class
